Cloud-Automation/Azure/LAB08-Virtual-Network-Setup/requirements.txt
azure-identity>=1.13.0
azure-mgmt-network>=25.1.0
azure-mgmt-resource>=23.0.1
azure-core>=1.26.4 